refactor embeddings.py so that it becomes a nice module.
