+++ src/diffusers/models/embeddings/__init__.py
@@ -1,0 +1,28 @@
+from .ip_adapter import (
+    IPAdapterFaceIDImageProjection,
+    IPAdapterFaceIDPlusImageProjection,
+    IPAdapterFullImageProjection,
+    IPAdapterPlusImageProjection,
+    IPAdapterPlusImageProjectionBlock,
+    MultiIPAdapterImageProjection,
+)
+from .misc import (
+    AttentionPooling,
+    CombinedTimestepLabelEmbeddings,
+    GaussianFourierProjection,
+    GLIGENTextBoundingboxProjection,
+    ImageHintTimeEmbedding,
+    ImagePositionalEmbeddings,
+    ImageProjection,
+    ImageTimeEmbedding,
+    LabelEmbedding,
+    PatchEmbed,
+    PixArtAlphaCombinedTimestepSizeEmbeddings,
+    PixArtAlphaTextProjection,
+    SinusoidalPositionalEmbedding,
+    TextImageProjection,
+    TextImageTimeEmbedding,
+    TextTimeEmbedding,
+    TimestepEmbedding,
+    Timesteps,
+)
